requirements.txt
apache-airflow==2.10.0
apache-airflow-providers-postgres==5.12.0
airflow-providers-clickhouse==0.0.1
airflow-clickhouse-plugin[common.sql]==1.4.0
airflow-clickhouse-plugin==1.4.0
dbt==1.0.0.38.12
dbt-core==1.8.6
dbt-postgres==1.8.2
dbt-clickhouse==1.8.0
csvkit==1.0.7
pandas==2.0.3
python-stdnum==1.17
deep-translator==1.9.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.5
bs4==0.0.1
beautifulsoup4==4.11.1
numpy==1.24.4
requests==2.32.3
python-dotenv==1.0.0
clickhouse_connect==0.6.22
loguru==0.7.2
pendulum==3.0.0